separating remaining units
